Move duplicate code of playPauseBtn to `CustomExoPlayerView`
